v20.0.0: Type uint64 fields as uint64 (BigInt-safety fix)

+++ src/api/chia/harvester/harvester.ts
@@ -1,6 +1,6 @@
 import { bytes, Optional, str } from "../types/_python_types_";
 import { G1Element } from "../../chia_rs/chia-bls/lib";
-import { int } from "../../chia_rs/wheel/python/sized_ints";
+import { int, uint64 } from "../../chia_rs/wheel/python/sized_ints";
 import { bytes32 } from "../../chia_rs/wheel/python/sized_bytes";
 
 export type Plot = {
@@ -10,7 +10,7 @@ export type Plot = {
   pool_public_key: Optional<G1Element>;
   pool_contract_puzzle_hash: Optional<bytes32>;
   plot_public_key: G1Element;
-  file_size: int;
-  time_modified: int;
+  file_size: uint64;
+  time_modified: uint64;
   compression_level: int;
 };
